+++ framework/src/main/java/org/checkerframework/framework/stub/AnnotationFileElementTypes.java
@@ -391,6 +391,8 @@ public class AnnotationFileElementTypes {
      *     does not appear in an annotation file.
      */
     public Set<AnnotationMirror> getDeclAnnotations(Element elt) {
+        // If currently parsing a file, return an empty set.
+        // The only exception is package because we always load package-info eagerly.
         if (parsing && elt.getKind() != ElementKind.PACKAGE) {
             return Collections.emptySet();
         }