+++ checker/tests/nullness-stubfile/EisopIssue270.java
@@ -1,0 +1,7 @@
+import java.util.Set;
+
+public class EisopIssue270 {
+    void foo(Set<Object> so, Set<? extends Object> seo) {
+        so.retainAll(seo);
+    }
+}