+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java
@@ -4180,6 +4180,8 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
         }
 
         // If parsing annotation files, return only the annotations in the element.
+        // The only exception is package because we always load package-info eagerly
+        // and there is no parent element to parse.
         boolean isParsing =
                 stubTypes.isParsing()
                         || ajavaTypes.isParsing()