+++ checker/src/main/java/org/checkerframework/checker/nullness/NullnessVisitor.java
@@ -721,14 +721,9 @@ public class NullnessVisitor
 
     @Override
     public Void visitNewClass(NewClassTree node, Void p) {
-        Tree enclosingExpr = node.getEnclosingExpression();
+        ExpressionTree enclosingExpr = node.getEnclosingExpression();
         if (enclosingExpr != null) {
-            AnnotatedDeclaredType type =
-                    (AnnotatedDeclaredType) atypeFactory.getAnnotatedType(enclosingExpr);
-            AnnotationMirror nullableAnno = type.getAnnotation(Nullable.class);
-            if (nullableAnno != null) {
-                checker.reportError(node, "nullness.on.receiver");
-            }
+            checkForNullability(enclosingExpr, "nullness.on.receiver");
         }
         AnnotatedDeclaredType type = atypeFactory.getAnnotatedType(node);
         ExpressionTree identifier = node.getIdentifier();