+++ docs/CHANGELOG.md
@@ -7,7 +7,8 @@ Version 3.22.1-eisop1 (June ?, 2022)
 
 **Closed issues:**
 
-typetools#2722, typetools#2995, typetools#3015, typetools#3025, typetools#3030.
+typetools#2722, typetools#2995, typetools#3015, typetools#3025, typetools#3027,
+typetools#3030.
 
 
 Version 3.22.1 (June 1, 2022)