+++ docs/CHANGELOG.md
@@ -21,7 +21,7 @@ memory consumption.
 
 **Closed issues:**
 
-eisop#270, typetools#5189.
+eisop#270, typetools#5189, eisop#281.
 
 
 Version 3.22.2 (July 1, 2022)